+++ other-simple-microservice/src/main/java/io/github/ddongeee/day2/Day2Application.java
@@ -1,0 +1,11 @@
+package io.github.ddongeee.day2;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Day2Application {
+	public static void main(String[] args) {
+		SpringApplication.run(Day2Application.class, args);
+	}
+}